Store positions/velocity/height as float32 (schema v33) (#90)
* Store positions/velocity/height as float32 (schema v33)

The MVD wire carries floating-point coordinates (eighth-unit fixed point,
or true floats under the float-coords extension) that mvd-reader already
decodes as float32, but the analytics pipeline truncated each player
origin to int32 in PositionTrack — losing up to ~1 unit per axis and,
worse, deriving the v32 velocity from those rounded positions.

PositionTrack X/Y/Z, VX/VY/VZ and H are now float32, as are AirgibEvent
heights and the view-layer Position3D/Velocity3D/Hgt and bucket/columnar
position/velocity/height outputs. Velocity is recomputed by central
difference over the full-resolution positions, so the old +/-1-unit
quantization noise is gone.

Values are kept at native float32 resolution in memory; the new
result.Coord / result.Coords types round only the JSON text to 3
decimals (lossless for eighth-unit positions, shedding the float
division/epsilon tail on derived velocity and height). The
PositionTrack.H NoFloor sentinel changes from math.MinInt32 to -1e9
(-1000000000), which float32 represents exactly so it round-trips through
JSON (math.MinInt32 serialized as a float32 became -2147483600).

Bump CurrentSchemaVersion to 33 and regenerate the golden corpus. Time
axes stay int32 ms; view angles stay int16 raw angle16. Docs updated
(RESULT_SCHEMA.md, API.md, READMEs).

* Round coords in JSON via MarshalJSON, keep float32 fields native

Follow-up to the float32 migration: keep the data-model fields plain
float32 (no named type rippling through the analyzer) and do the
3-decimal rounding only at serialization.

PositionTrack X/Y/Z/H/VX/VY/VZ go back to []float32; a container
PositionTrack.MarshalJSON rounds the coordinate columns to 3 decimals in
the JSON text via a shadow struct (the int columns and omitempty are left
to the standard encoder). This covers the dense outputs — the full result
/ goldens and every /stream-slice sub-track (all *PositionTrack). The
view buckets keep the result.Coord/Coords value types for their
map[string]any / columnar outputs, which a container method can't reach.

state-at (Position3D/Velocity3D/Hgt) and AirgibEvent heights revert to
plain float32 emitted at full precision — low-volume surfaces that don't
need rounding. Goldens regenerated: only airgib height values change
(back to full precision); PositionTrack output is byte-identical to the
prior rounded form. Docs updated.

* Round airgib table heights to 1 decimal

Airgib height / height-above-shooter are now full-precision float32 on
the wire (e.g. 366.21503). Round them to 1 decimal for display in the
Key Moments airgib table; sorting still uses the raw values.

+++ mvd-analytics/analyzer/airgibs.go
@@ -107,7 +107,7 @@ func airgibsPost(res *Result, co *CoreOutputs) {
 		// samples to the hit; 0 when the shooter has no sample close
 		// enough (and on a genuine dead-level hit — omitempty folds
 		// both, the neutral value either way).
-		dz := int32(0)
+		dz := float32(0)
 		if as := streamByName[d.Attacker]; as != nil && as.Position != nil && len(as.Position.T) > 0 {
 			ai := nearestSampleIndex(as.Position.T, d.Time)
 			if ai >= 0 && absI32(as.Position.T[ai]-d.Time) <= airgibPosMaxGapMs {
